+++ src/constants/MuiThemeOptions.ts
@@ -8,13 +8,24 @@ import SilkaBold from '../assets/fonts/silka/silka-bold-webfont.ttf';
 import SilkaBlack from '../assets/fonts/silka/silka-black-webfont.ttf';
 import { ThemeOptions } from '@mui/material';
 
+declare module '@mui/material/styles' {
+  interface BreakpointOverrides {
+    xs: false;
+    sm: false;
+    md: false;
+    lg: false;
+    xl: false;
+    mobile: true;
+    tablet: true;
+    laptop: true;
+    desktop: true;
+  }
+}
+
 export const MUI_THEME_OPTIONS: ThemeOptions = {
   typography: {
     fontFamily: ['Silka', 'sans-serif', 'Arial'].join(','),
     fontSize: 14,
-    h1: {
-      fontSize: '2.5rem',
-    },
   },
   components: {
     MuiCssBaseline: {
@@ -30,6 +41,34 @@ export const MUI_THEME_OPTIONS: ThemeOptions = {
           { fontFamily: 'Silka', src: `url(${SilkaBlack})`, fontWeight: 900 },
         ],
       },
+    },
+    MuiContainer: {
+      defaultProps: {
+        fixed: true,
+      },
+      styleOverrides: {
+        root: {
+          minHeight: '100vh',
+        },
+      },
+    },
+    MuiToggleButtonGroup: {
+      defaultProps: {
+        size: 'small',
+      },
+    },
+    MuiTextField: {
+      defaultProps: {
+        size: 'small',
+      },
+    },
+  },
+  breakpoints: {
+    values: {
+      mobile: 0,
+      tablet: 640,
+      laptop: 1024,
+      desktop: 1200,
     },
   },
 };