+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## 29/11/2022
 - Added Mui as styling framework.
-- Added custom font family.
+- Added `Silka` as custom font family.
 - Added dark mode.
+- Added routing for `index` and `404` pages.